requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
jinja2==3.1.4
sqlalchemy==2.0.31
pydantic==2.7.4
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
cryptography==42.0.8
httpx==0.27.0
python-multipart==0.0.9
